Finally done making new readme, add changesets, adjust types. Finally can make a pull request
