+++ examples/react/postprocessing/CHANGELOG.md
@@ -1,0 +1,8 @@
+# postprocessing
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - glsl-pipeline@2.0.0
