+++ examples/react/shader_options/package.json
@@ -1,7 +1,7 @@
 {
   "name": "shader_options",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.0.1",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@react-three/fiber": "^8.15.11",
-    "glsl-pipeline": "^1.0.6",
+    "glsl-pipeline": "^2.0.0",
     "leva": "^0.9.35",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
