+++ examples/react/typescript_postprocessing/CHANGELOG.md
@@ -1,0 +1,8 @@
+# typescript_postprocessing
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - glsl-pipeline@2.0.0
