+++ examples/react/typescript_postprocessing/package.json
@@ -1,7 +1,7 @@
 {
   "name": "typescript_postprocessing",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.0.1",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@react-three/fiber": "^8.15.11",
-    "glsl-pipeline": "^1.0.6",
+    "glsl-pipeline": "^2.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "resolve-lygia": "^1.0.4",
