+++ examples/vanilla/2d_clock/CHANGELOG.md
@@ -1,0 +1,8 @@
+# 2d_clock
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - glsl-pipeline@2.0.0
