+++ examples/vanilla/3d_pbr_shadow/CHANGELOG.md
@@ -1,0 +1,8 @@
+# 3d_pbr_shadow
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - glsl-pipeline@2.0.0
