+++ examples/vanilla/3d_pbr_shadow_dof/CHANGELOG.md
@@ -1,0 +1,8 @@
+# 3d_pbr_shadow_dof
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - glsl-pipeline@2.0.0
