+++ examples/vanilla/3d_trails/CHANGELOG.md
@@ -1,0 +1,8 @@
+# 3d_trails
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - glsl-pipeline@2.0.0
