--- package/.gitignore
@@ -1,3 +1,0 @@
-package-lock.json
-node_modules
-.DS_Store
