--- package/.npmignore
@@ -1,14 +1,0 @@
-node_modules/
-example/
-dist/
-build/
-Desktop.ini
-$RECYCLE.BIN/
-.DS_Store
-.vscode
-.docz/
-package-lock.json
-yarn-error.log
-pnpm-debug.log
-.parcel-cache
-pnpm-lock.yaml
