+++ package/CHANGELOG.md
@@ -1,0 +1,7 @@
+# glsl-pipeline
+
+## 2.0.0
+
+### Major Changes
+
+- New GlslPipeline Version that supports react & typescript!
