+++ package/src/types-helper/index.ts
@@ -1,1 +1,2 @@
 export * from "./ts-utils"
+export * from './ts-main'
